+++ src/linux/impl_window.rs
@@ -3,7 +3,7 @@ use std::str;
 use xcb::{
     x::{
         Atom, Drawable, GetGeometry, GetProperty, GetPropertyReply, InternAtom, QueryPointer,
-        TranslateCoordinates, Window, ATOM_ATOM, ATOM_NONE, ATOM_STRING, ATOM_WINDOW,
+        TranslateCoordinates, Window, ATOM_ATOM, ATOM_CARDINAL, ATOM_NONE, ATOM_STRING,
         ATOM_WM_CLASS, ATOM_WM_NAME,
     },
     Connection, Xid,
@@ -19,14 +19,15 @@ pub(crate) struct ImplWindow {
     pub id: u32,
     pub title: String,
     pub app_name: String,
+    pub pid: u32,
     pub current_monitor: ImplMonitor,
     pub x: i32,
     pub y: i32,
+    pub z: i32,
     pub width: u32,
     pub height: u32,
     pub is_minimized: bool,
     pub is_maximized: bool,
-    pub is_focused: bool,
 }
 
 fn get_atom(conn: &Connection, name: &str) -> XCapResult<Atom> {
@@ -66,25 +67,24 @@ fn get_window_property(
     Ok(window_property_reply)
 }
 
-fn get_focused_window(conn: &Connection, root_window: Window) -> XCapResult<Window> {
-    let active_window_atom = get_atom(conn, "_NET_ACTIVE_WINDOW")?;
+pub fn get_window_pid(conn: &Connection, window: &Window) -> XCapResult<u32> {
+    let wm_pid_atom = get_atom(conn, "_NET_WM_PID")?;
 
-    let active_window_reply =
-        get_window_property(conn, root_window, active_window_atom, ATOM_WINDOW, 0, 1)?;
+    let reply = get_window_property(conn, *window, wm_pid_atom, ATOM_CARDINAL, 0, 4)?;
+    let value = reply.value::<u32>();
 
-    let active_window = active_window_reply
-        .value::<Window>()
+    value
         .first()
+        .ok_or(XCapError::new("Get window pid failed"))
         .copied()
-        .unwrap_or(Window::none());
-
-    Ok(active_window)
 }
 
 impl ImplWindow {
     fn new(
         conn: &Connection,
         window: &Window,
+        pid: u32,
+        z: i32,
         impl_monitors: &Vec<ImplMonitor>,
     ) -> XCapResult<ImplWindow> {
         let title = {
@@ -183,29 +183,20 @@ impl ImplWindow {
             )
         };
 
-        let is_focused = {
-            let setup = conn.get_setup();
-            let screen = setup
-                .roots()
-                .next()
-                .ok_or(XCapError::new("No screen found"))?;
-            let focused_window = get_focused_window(conn, screen.root())?;
-            focused_window == *window
-        };
-
         Ok(ImplWindow {
             window: *window,
             id: window.resource_id(),
             title,
             app_name,
+            pid,
             current_monitor,
             x,
             y,
+            z,
             width,
             height,
             is_minimized,
             is_maximized,
-            is_focused,
         })
     }
 
@@ -214,11 +205,14 @@ impl ImplWindow {
         let setup = conn.get_setup();
 
         // https://github.com/rust-x-bindings/rust-xcb/blob/main/examples/get_all_windows.rs
-        let client_list_atom = get_atom(&conn, "_NET_CLIENT_LIST")?;
+        // https://specifications.freedesktop.org/wm-spec/1.5/ar01s03.html#id-1.4.4
+        // list all windows by stacking order
+        let client_list_atom = get_atom(&conn, "_NET_CLIENT_LIST_STACKING")?;
 
         let mut impl_windows = Vec::new();
         let impl_monitors = ImplMonitor::all()?;
 
+        let mut z = -1;
         for screen in setup.roots() {
             let root_window = screen.root();
 
@@ -237,14 +231,24 @@ impl ImplWindow {
                     client_list_atom,
                     ATOM_NONE,
                     0,
-                    100,
+                    1024,
                 ) {
                     Ok(list_window_reply) => list_window_reply,
                     _ => continue,
                 };
 
                 for client in list_window_reply.value::<Window>() {
-                    if let Ok(impl_window) = ImplWindow::new(&conn, client, &impl_monitors) {
+                    z += 1;
+                    let pid = match get_window_pid(&conn, client) {
+                        Ok(pid) => pid,
+                        err => {
+                            log::error!("{:?}", err);
+                            continue;
+                        }
+                    };
+
+                    if let Ok(impl_window) = ImplWindow::new(&conn, client, pid, z, &impl_monitors)
+                    {
                         impl_windows.push(impl_window);
                     } else {
                         log::error!(
@@ -257,31 +261,13 @@ impl ImplWindow {
             }
         }
 
+        impl_windows.reverse();
+
         Ok(impl_windows)
     }
 }
 
 impl ImplWindow {
-    pub fn refresh(&mut self) -> XCapResult<()> {
-        let (conn, _) = Connection::connect(None)?;
-        let impl_monitors = ImplMonitor::all()?;
-        let impl_window = ImplWindow::new(&conn, &self.window, &impl_monitors)?;
-
-        self.window = impl_window.window;
-        self.id = impl_window.id;
-        self.title = impl_window.title;
-        self.app_name = impl_window.app_name;
-        self.current_monitor = impl_window.current_monitor;
-        self.x = impl_window.x;
-        self.y = impl_window.y;
-        self.width = impl_window.width;
-        self.height = impl_window.height;
-        self.is_minimized = impl_window.is_minimized;
-        self.is_maximized = impl_window.is_maximized;
-        self.is_focused = impl_window.is_focused;
-
-        Ok(())
-    }
     pub fn capture_image(&self) -> XCapResult<RgbaImage> {
         capture_window(self)
     }
