+++ src/window.rs
@@ -14,6 +14,7 @@ impl Window {
 }
 
 impl Window {
+    /// List all windows, sorted by z coordinate.
     pub fn all() -> XCapResult<Vec<Window>> {
         let windows = ImplWindow::all()?
             .iter()
@@ -38,10 +39,9 @@ impl Window {
         &self.impl_window.title
     }
 
-    #[cfg(any(target_os = "macos", target_os = "windows"))]
     /// The window process id
-    pub fn process_id(&self) -> u32 {
-        self.impl_window.process_id
+    pub fn pid(&self) -> u32 {
+        self.impl_window.pid
     }
     /// The window current monitor
     pub fn current_monitor(&self) -> Monitor {
@@ -55,7 +55,6 @@ impl Window {
     pub fn y(&self) -> i32 {
         self.impl_window.y
     }
-    #[cfg(any(target_os = "macos", target_os = "windows"))]
     /// The window z coordinate.
     pub fn z(&self) -> i32 {
         self.impl_window.z
@@ -79,9 +78,6 @@ impl Window {
 }
 
 impl Window {
-    pub fn refresh(&mut self) -> XCapResult<()> {
-        self.impl_window.refresh()
-    }
     pub fn capture_image(&self) -> XCapResult<RgbaImage> {
         self.impl_window.capture_image()
     }
